Intializing From an Open Scene
Updated scenefile.py to allow a path to  be intialized from an open scene. It will take that scene's path as its own if a path isn't already specified in SceneFile(). If the scene is a new scene, it will give a warning saying it is unable to intialize from a new scene.