Finished Updating UI
Added two labels for the object to scatter to and the object to scatter with. I also added a button that takes the currently selected objects and fills in the labels with their names. I also made it where the window doesn't fully open unless only two items are selected (might change this later).